# docs/api/inventory-service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartRetailX - Inventory Service",
    "description": "Inventory service with stock reservation and oversell prevention",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Report healthy only when the inventory table is reachable.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/{product_id}": {
      "get": {
        "summary": "Get Stock",
        "description": "Get the current stock level for a product.",
        "operationId": "get_stock_api_v1_inventory__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/admin/low-stock": {
      "get": {
        "summary": "Low Stock",
        "description": "Every product at or below `threshold` available units \u2014 the admin\ndashboard's low-stock alert list. A flat, global threshold by design\n(see repository.list_low_stock()'s own docstring for the \"something\nsmarter\" alternatives considered and not built).",
        "operationId": "low_stock_api_v1_inventory_admin_low_stock_get",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 10,
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItem"
                  },
                  "title": "Response Low Stock Api V1 Inventory Admin Low Stock Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/admin/batch": {
      "post": {
        "summary": "Get Stock Batch",
        "description": "Return current stock for many admin-catalogue rows in one request.",
        "operationId": "get_stock_batch_api_v1_inventory_admin_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InventoryItem"
                  },
                  "type": "array",
                  "title": "Response Get Stock Batch Api V1 Inventory Admin Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/{product_id}/reserve": {
      "post": {
        "summary": "Reserve Stock",
        "description": "Reserve stock for a product. Returns 409 if insufficient stock.\n\nNot called by the Order saga, which reserves a whole basket at once via\nthe all-or-nothing /reserve batch endpoint below. This single-item form\nexists for manual admin stock management (e.g. holding stock for an\noffline sale), so it is admin-only rather than customer- or\nsaga-reachable.",
        "operationId": "reserve_stock_api_v1_inventory__product_id__reserve_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quantity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/{product_id}/release": {
      "post": {
        "summary": "Release Stock",
        "description": "Release previously reserved stock for a product. Admin-only (see reserve_stock).",
        "operationId": "release_stock_api_v1_inventory__product_id__release_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quantity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/{product_id}/confirm": {
      "post": {
        "summary": "Confirm Stock",
        "description": "Confirm a sale \u2014 permanently remove reserved stock. Admin-only (see reserve_stock).",
        "operationId": "confirm_stock_api_v1_inventory__product_id__confirm_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quantity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/{product_id}/add": {
      "post": {
        "summary": "Add Stock",
        "description": "Add stock (restock or create initial record).",
        "operationId": "add_stock_api_v1_inventory__product_id__add_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quantity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/reserve": {
      "post": {
        "summary": "Reserve Batch",
        "operationId": "reserve_batch_api_v1_inventory_reserve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/StockOperation"
                },
                "type": "array",
                "title": "Items"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/release": {
      "post": {
        "summary": "Release Batch",
        "operationId": "release_batch_api_v1_inventory_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/StockOperation"
                },
                "type": "array",
                "title": "Items"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/confirm": {
      "post": {
        "summary": "Confirm Batch",
        "operationId": "confirm_batch_api_v1_inventory_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/StockOperation"
                },
                "type": "array",
                "title": "Items"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "available_quantity": {
            "type": "integer",
            "title": "Available Quantity"
          },
          "reserved_quantity": {
            "type": "integer",
            "title": "Reserved Quantity",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "available_quantity"
        ],
        "title": "InventoryItem"
      },
      "StockBatchRequest": {
        "properties": {
          "product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Product Ids"
          }
        },
        "type": "object",
        "required": [
          "product_ids"
        ],
        "title": "StockBatchRequest",
        "description": "Product ids whose current stock levels should be returned together."
      },
      "StockOperation": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "StockOperation",
        "description": "One line of a batch reserve / release / confirm."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}